+++ src/utils/decorators/input-label/__spec__.js
@@ -148,6 +148,32 @@ describe('InputLabel', () => {
     instanceHelp = TestUtils.renderIntoDocument(<ExtendedClassEight labelHelp='Some Helpful Content'/>);
   });
 
+  describe('labelClasses', () => {
+    describe('when inline', () => {
+      it('adds the relevant class', () => {
+        let ExtendedClassOne = InputLabel(HelpClass);
+        let instance = TestUtils.renderIntoDocument(<ExtendedClassOne labelInline={ true } />);
+        expect(instance.labelClasses).toEqual('common-input__label common-input__label--inline');
+      });
+    });
+
+    describe('when help is enabled', () => {
+      it('adds the relevant class', () => {
+        let ExtendedClassOne = InputLabel(HelpClass);
+        let instance = TestUtils.renderIntoDocument(<ExtendedClassOne fieldHelp={ true } />);
+        expect(instance.labelClasses).toEqual('common-input__label common-input__label--help');
+      });
+    });
+
+    describe('when align right', () => {
+      it('adds the relevant class', () => {
+        let ExtendedClassOne = InputLabel(HelpClass);
+        let instance = TestUtils.renderIntoDocument(<ExtendedClassOne labelAlign={ true } />);
+        expect(instance.labelClasses).toEqual('common-input__label common-input__label--align-right');
+      });
+    });
+  });
+
   describe('labelHTML', () => {
     describe('when label is set to false', () => {
       it('does not add a label', () => {