+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ## Misc
 
 * Added 'small' button option - renders pill-like secondary button.
+* Inline labels can now be aligned right.
 * Made portrait inline-block to allow label to sit inline. 
 
 # 0.8.1